[refactor]: modularize
