+++ src/app/globals.css
@@ -1,4 +1,4 @@
-@import "./fonts/fonts.css";
+@import "../fonts/fonts.css";
 @import url("https://fonts.googleapis.com/css2?family=EB+Garamond:ital,wght@0,400..800;1,400..800&family=Tiro+Bangla:ital@0;1&display=swap");
 
 @tailwind base;
