+++ src/app/kigotumi/page.tsx
@@ -1,4 +1,4 @@
-import { Messages } from "../screens";
+import { Messages } from "../../screens";
 
 export default function KigoTumi() {
   return (
