+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import { AudioProvider, DarkModeProvider } from "./context";
-import { Layout, Loader } from "./components";
+import { AudioProvider, DarkModeProvider } from "../context";
+import { Layout, Loader } from "../components";
 import { CSSProperties } from "react";
 
 const inter = Inter({ subsets: ["latin"] });
