+++ src/components/Flashlight.tsx
