+++ src/components/Spots.tsx
